+++ src/flag_reading.c
@@ -1,0 +1,99 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   flag_reading.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cdahlhof <cdahlhof@student.42.fr>          #+#  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025-08-05 19:17:07 by cdahlhof          #+#    #+#             */
+/*   Updated: 2025-08-05 19:17:07 by cdahlhof         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../inc/ft_printf.h"
+
+void	read_flags(void)
+{
+	t_data	*data;
+
+	data = *query();
+	if (data->format[data->i + data->len] == ' ')
+		data->do_sign = true;
+	if (data->format[data->i + data->len] == '+')
+		data->sign_positive = true;
+	if (data->format[data->i + data->len] == '-')
+		data->leftbound = true;
+	if (data->format[data->i + data->len] == '#')
+		data->pointer_prefix = true;
+	if (data->format[data->i + data->len] == '0')
+		data->padd_char = '0';
+}
+
+void	read_numbers(void)
+{
+	t_data	*data;
+
+	data = *query();
+	if (data->format[data->i + data->len] == '.')
+	{
+		data->precision = ft_atoi(&data->format[data->i + data->len + 1]);
+		if (data->precision != 0)
+			data->add += ft_log(data->precision, 10);
+		if (data->format[data->i + data->len + 1] == '0')
+			data->add++;
+		data->do_precision = true;
+	}
+	if (ft_strchr(data->format[data->i + data->len], "123456789"))
+	{
+		data->do_width = true;
+		data->width = ft_atoi(&data->format[data->i + data->len]);
+		data->add = ft_log(data->width, 10);
+	}
+}
+
+int	delegate_handlers(va_list a)
+{
+	t_data	*data;
+
+	data = *query();
+	if (data->insert_identifier == 'i' || data->insert_identifier == 'd')
+		return (handle_d_i(a));
+	if (data->insert_identifier == 'u')
+		return (handle_u(a));
+	if (data->insert_identifier == 'x')
+		return (handle_x(a));
+	if (data->insert_identifier == 'p')
+		return (handle_p(a));
+	if (data->insert_identifier == 'X')
+		return (handle_big_x(a));
+	if (data->insert_identifier == 'c')
+		return (handle_c(a));
+	if (data->insert_identifier == '%')
+		return (handle_percent());
+	if (data->insert_identifier == 's')
+		return (handle_s(a));
+	return (0);
+}
+
+int	trigger_insert(va_list a)
+{
+	t_data	*data;
+
+	data = *query();
+	data->len = 0;
+	while (data->format[data->i + data->len] && \
+					!ft_strchr(data->format[data->i + data->len], "discuxXp%"))
+	{
+		data->add = 1;
+		read_flags();
+		read_numbers();
+		data->len += data->add;
+	}
+	if (data->format[data->i + data->len])
+		data->insert_identifier = \
+					*ft_strchr(data->format[data->i + data->len], "discuxXp%");
+	data->skipped = data->len + 1;
+	delegate_handlers(a);
+	reset_flags();
+	return (0);
+}
