+++ src/mem_utils.c
@@ -1,0 +1,100 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   mem_utils.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: cdahlhof <cdahlhof@student.42.fr>          #+#  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025-08-05 19:09:28 by cdahlhof          #+#    #+#             */
+/*   Updated: 2025-08-05 19:09:28 by cdahlhof         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../inc/ft_printf.h"
+
+void	ft_bzero(void *s, size_t n)
+{
+	unsigned char	*c;
+	size_t			i;
+
+	c = (unsigned char *)s;
+	i = 0;
+	while (n > i)
+	{
+		c[i] = '\0';
+		i++;
+	}
+}
+
+void	ft_memset(void *s, char r, size_t n)
+{
+	unsigned char	*c;
+	size_t			i;
+
+	c = (unsigned char *)s;
+	i = 0;
+	while (n > i)
+	{
+		c[i] = r;
+		i++;
+	}
+}
+
+void	*ft_memcpy(void *dest, const void *src, size_t n)
+{
+	size_t			i;
+	unsigned char	*s;
+	unsigned char	*d;
+
+	i = 0;
+	if (!dest && !src)
+		return (NULL);
+	s = (unsigned char *)src;
+	d = (unsigned char *)dest;
+	while (n > i)
+	{
+		d[i] = s[i];
+		i++;
+	}
+	return (dest);
+}
+
+int	ft_atoi(const char *nptr)
+{
+	int	i;
+	int	sign;
+	int	num;
+
+	i = 0;
+	sign = 1;
+	num = 0;
+	while ((nptr[i] > 8 && nptr[i] < 14) || nptr[i] == 32)
+		i++;
+	if (nptr[i] == 43 || nptr[i] == 45)
+	{
+		if (nptr[i] == 45)
+			sign *= -1;
+		i++;
+	}
+	while (nptr[i] > 47 && nptr[i] < 58)
+	{
+		num = num * 10 + nptr[i] - '0';
+		i++;
+	}
+	return (num * sign);
+}
+
+int	ft_log(unsigned long num, int base)
+{
+	int	digits;
+
+	digits = 0;
+	if (num <= 0)
+		digits++;
+	while (num)
+	{
+		digits++;
+		num = num / base;
+	}
+	return (digits);
+}
